+++ CHANGELOG.md
@@ -20,6 +20,7 @@ different versioning scheme, following the Haskell community's
 ### C++ ###
 * Fixed a crash in `bond::ext::grpc::server`'s destructor when invoked on
   a moved-from object instance.
+* Added move-assignment operator to `bond::blob`
 
 ## 8.0.0: 2018-05-30 ##
 * `gbc` & compiler library: 0.11.0.0